Move is_running to testing, poll for service state in CI
- Remove is_running() from client.py (only try_connect remains)
- Add supervisor_is_running() test helper to testing/supervisor_cleanup
- Replace fixed sleep with polling loop in test_add_and_check_service
  (CI machines are slower, 0.5s wasn't enough)
