+++ tools/claude_hooks/testing/BUILD.bazel
@@ -46,6 +46,10 @@ py_library(
     srcs = ["supervisor_cleanup.py"],
     imports = ["../../.."],
     visibility = ["//tools/claude_hooks:__subpackages__"],
+    deps = [
+        "//tools/claude_hooks:settings",
+        "//tools/claude_hooks/supervisor:client",
+    ],
 )
 
 py_library(
