+++ tools/claude_hooks/testing/supervisor_cleanup.py
@@ -1,4 +1,4 @@
-"""Supervisor cleanup utilities for testing."""
+"""Supervisor cleanup and assertion utilities for testing."""
 
 from __future__ import annotations
 
@@ -8,6 +8,14 @@ import signal
 import time
 from collections.abc import Generator
 from pathlib import Path
+
+from tools.claude_hooks.settings import HookSettings
+from tools.claude_hooks.supervisor.client import try_connect
+
+
+async def supervisor_is_running(settings: HookSettings) -> bool:
+    """Check if supervisord is running (test helper)."""
+    return await try_connect(settings) is not None
 
 
 def stop_supervisor_by_pidfile(pidfile: Path) -> None:
